Python: move concurrency tests into existing suite
